Extracted article types from ArticleService to types/Article
